increase default chunkSize to 30 MB
This drastically improves performance on faster connections, at the cost
of potentially needing to retry a larger chunk of data when a connection
is interrupted.

It's impossible to serve both fast and slow/unreliable internet
connections well with a single fixed setting here, so in the future this
should be made to scale dynamically to 1) maximize upload speed on fast
connections, and 2) constrain the amount of time that will be wasted
re-uploading a partial segment if it is interrupted.